rgw/radosgw-admin: extract command types into a header
Move SimpleCmd and the OPT command enum from radosgw-admin.cc into
radosgw-admin.h so later translation units under radosgw-admin/ can
share them without depending on the rest of the monolith.

Keep command tables, usage text, and handler implementation in the .cc.
Qualify SimpleCmd with std:: types for header hygiene.
